Added Getters and Setters + Added a test sample
Added getters and setters for the private variables in the Recipe class.

Made an array list of type <Recipe> to add multiple recipes.

Created a test sample for more efficient testing, so I don't need to answer the prompt.

I added a while (application) to keep the application running until the user wishes to close it. (It is all commented out for now to just simply test methods I am creating in the Recipe class.