Updated Input prompts + Fixed Scaling issue
I firstly went through the user prompts and made it fairly clean. Allowing the user to have multiple options to choose from. Such as; adding a new recipe, display a certain recipe, deleting a recipe and etc.

I was then experiencing an issue with the Scaling method because when I was scaling a recipe, it was setting the original recipe to the newly scaled one. Which is not what I wanted the method to do. I did some reading online and research, and then some trouble shooting, I then realized by attempting to create a copy of the ingredients array (the private variable) that I was actually creating a reference point to the private variable where the original ingredients were stored, and using the Scale() was altering the data stored in there. I fixed that by using the .Clone().

I also learnt how to efficiently populate the private variables of type arrays. I wasn't sure how to initialize a 2D array of a private variable from the main, because I didn't know that I had to include the size of the array when initializing it.